+++ src/app/services/contest-voting-card-layout.service.ts
@@ -43,7 +43,14 @@ export class ContestVotingCardLayoutService {
 
   public async setLayout(contestId: string, layout: ContestVotingCardLayout): Promise<void> {
     await firstValueFrom(
-      this.client.setLayout(new SetContestVotingCardLayoutRequest({ ...layout, templateId: layout.template!.id, contestId })),
+      this.client.setLayout(
+        new SetContestVotingCardLayoutRequest({
+          ...layout,
+          contestId,
+          templateId: layout.template!.id,
+          color: layout.color,
+        }),
+      ),
     );
   }
 }