+++ src/app/services/domain-of-influence-voting-card-layout.service.ts
@@ -89,6 +89,7 @@ export class DomainOfInfluenceVotingCardLayoutService {
           ...layout,
           domainOfInfluenceId,
           templateId: (layout.domainOfInfluenceTemplate?.id ?? 0 > 0) ? { value: layout.domainOfInfluenceTemplate!.id } : undefined,
+          color: layout.domainOfInfluenceColor,
         }),
       ),
     );
@@ -101,6 +102,7 @@ export class DomainOfInfluenceVotingCardLayoutService {
           ...layout,
           domainOfInfluenceId,
           templateId: (layout.overriddenTemplate?.id ?? 0 > 0) ? { value: layout.overriddenTemplate!.id } : undefined,
+          color: layout.overriddenColor,
         }),
       ),
     );